place seal and made 1 side of fold

+++ duet/scriptslibrary/Constants.cs
@@ -4,10 +4,10 @@ namespace StorybrewScripts
 {
     public class Constants {
         public static double beatLength = 714;
-        public static double xFloor = -107;
-        public static double xCeil = 747;
-        public static double width = 854;
-        public static double height = 480;
+        public static float xFloor = -107;
+        public static float xCeil = 747;
+        public static float width = 854;
+        public static float height = 480;
         public static int glowRadius = 5;
         public static string jpFont = "Yu Mincho";
         public static string enFont = "Adobe Garamond Pro";
